FIx up atomics AP. - Add proper documentation of the function declarations - Split into windows/gcc-clang/irq implementations.
