+++ core/platform/CMakeLists.txt
@@ -12,28 +12,18 @@ set(LF_PLATFORM_FILES
     lf_zephyr_clock_counter.c
     lf_zephyr_clock_kernel.c
     lf_rp2040_support.c
+    lf_atomic_windows.c
+    lf_atomic_gcc_clang.c
+    lf_atomic_irq.c
 )
 
-if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
-    set(LF_ATOMIC_FILE lf_atomic64.c)
-elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Darwin")
-    set(LF_ATOMIC_FILE lf_atomic64.c)
-elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Windows")
-    set(CMAKE_SYSTEM_VERSION 10.0)
-    message("Using Windows SDK version ${CMAKE_VS_WINDOWS_TARGET_PLATFORM_VERSION}")
-    set(LF_ATOMIC_FILE lf_atomic64.c)
-elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Nrf52")
+if(${CMAKE_SYSTEM_NAME} STREQUAL "Nrf52")
     list(APPEND REACTORC_COMPILE_DEFS PLATFORM_NRF52)
-    set(LF_ATOMIC_FILE lf_atomic32.c)
 elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Zephyr")
     list(APPEND REACTORC_COMPILE_DEFS PLATFORM_ZEPHYR)
-    set(LF_ATOMIC_FILE lf_atomic32.c)
     set(PLATFORM_ZEPHYR true)
 elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Rp2040")
-    set(LF_ATOMIC_FILE lf_atomic32.c)
     list(APPEND REACTORC_COMPILE_DEFS PLATFORM_RP2040)
-else()
-    message(FATAL_ERROR "Your platform is not supported! The C target supports Linux, MacOS and Windows.")
 endif()
 
 # Add the atomics implementation
